refactor: cleanup logs
