+++ tools/ui/src/lib/stores/chat.svelte.ts
@@ -23,11 +23,10 @@ import { mcpStore } from '$lib/stores/mcp.svelte';
 import { contextSize, isRouterMode } from '$lib/stores/server.svelte';
 import {
 	selectedModelName,
-	singleModelName,
 	modelsStore,
 	selectedModelContextSize
 } from '$lib/stores/models.svelte';
-import { formatParameters } from '$lib/utils/formatters';
+import { logResponseStats, logFlowSummary, debugTokenUsage } from '$lib/utils/token-audit';
 import {
 	normalizeModelName,
 	filterByLeafNodeId,
@@ -1199,7 +1198,7 @@ class ChatStore {
 				timings: ChatMessageTimings | undefined,
 				toolCalls: import('$lib/types/api').ApiChatCompletionToolCall[] | undefined
 			) => {
-				this.logResponseStats(timings);
+				logResponseStats(timings);
 				const updateData: Record<string, unknown> = {
 					content,
 					reasoningContent: reasoningContent || undefined,
@@ -1283,7 +1282,7 @@ class ChatStore {
 				return msg;
 			},
 			onFlowComplete: (finalTimings?: ChatMessageTimings) => {
-				this.logFlowSummary();
+				logFlowSummary();
 				if (finalTimings) {
 					const idx = conversationsStore.findMessageIndex(assistantMessage.id);
 
@@ -1408,8 +1407,8 @@ class ChatStore {
 					conversationsStore.updateMessageAtIndex(idx, uiUpdate);
 					await conversationsStore.updateCurrentNode(currentMessageId);
 					cleanupStreamingState();
-					this.logResponseStats(timings);
-					this.logFlowSummary();
+					logResponseStats(timings);
+					logFlowSummary();
 					if (onComplete) await onComplete(content);
 					if (isRouterMode()) modelsStore.fetchRouterModels().catch(console.error);
 
@@ -1948,8 +1947,8 @@ class ChatStore {
 
 						conversationsStore.updateConversationTimestamp();
 
-						this.logResponseStats(timings);
-						this.logFlowSummary();
+						logResponseStats(timings);
+						logFlowSummary();
 
 						this.setChatLoading(msg.convId, false);
 						this.clearChatStreaming(msg.convId);
@@ -2295,246 +2294,13 @@ class ChatStore {
 		};
 	}
 
-	private logResponseStats(timings: ChatMessageTimings | undefined): void {
-		const assistantMessages = (conversationsStore.activeMessages as DatabaseMessage[]).filter(
-			(m) => m.role === MessageRole.ASSISTANT
-		);
-		const index = assistantMessages.length;
-		const prompt = timings?.prompt_n ?? 0;
-		const predicted = timings?.predicted_n ?? 0;
-		console.log(`[ChatStore] response #${index}: ${prompt} reading / ${predicted} generation`);
-	}
-
-	private logFlowSummary(): void {
-		const messages = conversationsStore.activeMessages as DatabaseMessage[];
-		const assistantMessages = messages.filter((m) => m.role === MessageRole.ASSISTANT);
-		const agenticMessages = assistantMessages.filter(
-			(m) => m.timings?.agentic?.llm?.predicted_n != null
-		);
-
-		// Derive context from message data (activeProcessingState may be stale by now).
-		// KV in memory = sum across all committed assistant turns of
-		// (timings.prompt_n + timings.predicted_n). Cache hits are excluded
-		// because they reference tokens already counted by a prior turn's
-		// prompt_n - summing them would double-count.
-		let contextUsed = 0;
-		for (const msg of assistantMessages) {
-			if (!msg.timings) continue;
-			contextUsed += (msg.timings.prompt_n ?? 0) + (msg.timings.predicted_n ?? 0);
-		}
-		const contextTotal = modelsStore.getModelContextSize(singleModelName() ?? '');
-		const contextPercent =
-			contextTotal && contextTotal > 0 ? Math.round((contextUsed / contextTotal) * 100) : null;
-
-		console.group(
-			`[ChatStore] --- TOKEN AUDIT --- Context: ${formatParameters(contextUsed)} / ${contextTotal !== null ? formatParameters(contextTotal) : '—'} (${contextPercent !== null ? contextPercent + '%' : '?'})`
-		);
-
-		// For agentic flows, use agentic.llm (accumulated across turns).
-		// Base timings.predicted_n/prompt_n are overwritten by buildFinalTimings
-		// with the last turn's values, so per-message summation would double-count.
-		let totalRead = 0;
-		let totalOutput = 0;
-		let agg: { prompt_n?: number; predicted_n?: number } | undefined;
-
-		if (agenticMessages.length > 0) {
-			const lastAgentic = agenticMessages[agenticMessages.length - 1];
-			agg = lastAgentic.timings?.agentic?.llm;
-			totalRead = agg?.prompt_n ?? 0;
-			totalOutput = agg?.predicted_n ?? 0;
-
-			// Log each message for reference
-			assistantMessages.forEach((msg, idx) => {
-				const t = msg.timings;
-				const modelLabel = msg.model ? ` [${msg.model}]` : '';
-				if (t?.agentic?.llm?.predicted_n != null) {
-					console.log(`  #${idx + 1}${modelLabel}`);
-					console.log(`    Fresh:        ${t.agentic.llm.prompt_n ?? 0}`);
-					console.log(`    Generated:    ${t.agentic.llm.predicted_n ?? 0}`);
-				} else {
-					const fresh = t?.prompt_n ?? 0;
-					const cached = t?.cache_n ?? 0;
-					const totalInput = fresh + cached;
-					const generated = t?.predicted_n ?? 0;
-					console.log(`  #${idx + 1}${modelLabel}`);
-					console.log(`    Fresh:        ${fresh}`);
-					console.log(`    Cached:       ${cached}`);
-					console.log(`    Total input:  ${totalInput}`);
-					console.log(`    Generated:    ${generated}`);
-				}
-			});
-		} else {
-			assistantMessages.forEach((msg, idx) => {
-				const t = msg.timings;
-				const fresh = t?.prompt_n ?? 0;
-				const cached = t?.cache_n ?? 0;
-				const totalInput = fresh + cached;
-				const generated = t?.predicted_n ?? 0;
-				totalRead += totalInput;
-				totalOutput += generated;
-				const modelLabel = msg.model ? ` [${msg.model}]` : '';
-				console.log(`  #${idx + 1}${modelLabel}`);
-				console.log(`    Fresh:        ${fresh}`);
-				console.log(`    Cached:       ${cached}`);
-				console.log(`    Total input:  ${totalInput}`);
-				console.log(`    Generated:    ${generated}`);
-			});
-		}
-
-		// Current request values (what's in KV cache now)
-		const lastMsg = assistantMessages[assistantMessages.length - 1];
-		const currentRead =
-			agenticMessages.length > 0
-				? (agg?.prompt_n ?? 0)
-				: lastMsg?.timings
-					? (lastMsg.timings.prompt_n ?? 0) + (lastMsg.timings.cache_n ?? 0)
-					: 0;
-		const currentOutput =
-			agenticMessages.length > 0 ? (agg?.predicted_n ?? 0) : (lastMsg?.timings?.predicted_n ?? 0);
-
-		// Cumulative totals
-		console.log('%cCumulative (all responses):', 'font-weight: bold');
-		console.log(`  Total input: ${formatParameters(totalRead)} tok`);
-		console.log(`  Generated:   ${formatParameters(totalOutput)} tok`);
-
-		// Current (in KV cache)
-		console.log('%cCurrent (KV cache):', 'font-weight: bold');
-		console.log(`  Total input: ${formatParameters(currentRead)} tok`);
-		console.log(`  Generated:   ${formatParameters(currentOutput)} tok`);
-		console.log(`  ─────────────────────`);
-		console.log(`  KV total:    ${formatParameters(currentRead + currentOutput)} tok`);
-
-		console.groupEnd();
-	}
-
 	/**
 	 * Debug: full token usage audit for a conversation. Call from browser console:
 	 *   chatStore.debugTokenUsage()            // active conversation
 	 *   chatStore.debugTokenUsage('conv-id')   // specific conversation
 	 */
 	async debugTokenUsage(convIdOrUndefined?: string): Promise<void> {
-		let convId = convIdOrUndefined;
-		if (!convId) {
-			const active = conversationsStore.activeConversation;
-			if (!active) {
-				console.log('[Debug] No active conversation');
-				return;
-			}
-			convId = active.id;
-		}
-
-		const messages = await DatabaseService.getConversationMessages(convId as string);
-		const assistantMessages = messages.filter((m) => m.role === MessageRole.ASSISTANT);
-		const agenticMessages = assistantMessages.filter(
-			(m) => m.timings?.agentic?.llm?.predicted_n != null
-		);
-
-		// Derive context from message data. KV in memory = sum across all
-		// committed assistant turns of (timings.prompt_n + timings.predicted_n);
-		// cache hits are excluded because they reference tokens already
-		// counted by a prior turn's prompt_n.
-		let contextUsed = 0;
-		for (const msg of assistantMessages) {
-			if (!msg.timings) continue;
-			contextUsed += (msg.timings.prompt_n ?? 0) + (msg.timings.predicted_n ?? 0);
-		}
-		const contextTotal = modelsStore.getModelContextSize(singleModelName() ?? '');
-		const contextPercent =
-			contextTotal && contextTotal > 0 ? Math.round((contextUsed / contextTotal) * 100) : null;
-
-		console.group(
-			`[Debug] Token audit: ${convId} (${assistantMessages.length} responses) Context: ${formatParameters(contextUsed)} / ${contextTotal !== null ? formatParameters(contextTotal) : '—'} (${contextPercent !== null ? contextPercent + '%' : '?'})`
-		);
-
-		// Per-message breakdown
-		console.log('%cPer-message token usage:', 'font-weight: bold');
-		let totalRead = 0;
-		let totalOutput = 0;
-		let agg: { prompt_n?: number; predicted_n?: number } | undefined;
-
-		if (agenticMessages.length > 0) {
-			// Agentic: use accumulated values from the last agentic message
-			const lastAgentic = agenticMessages[agenticMessages.length - 1];
-			agg = lastAgentic.timings?.agentic?.llm;
-			totalRead = agg?.prompt_n ?? 0;
-			totalOutput = agg?.predicted_n ?? 0;
-
-			assistantMessages.forEach((msg, idx) => {
-				const t = msg.timings;
-				const modelLabel = msg.model ? ` [${msg.model}]` : '';
-				if (t?.agentic?.llm?.predicted_n != null) {
-					console.log(
-						`  #${idx + 1}: agentic (llm.predicted_n=${t.agentic.llm.predicted_n}, llm.prompt_n=${t.agentic.llm.prompt_n}) ${modelLabel}`
-					);
-				} else {
-					const read = (t?.prompt_n ?? 0) + (t?.cache_n ?? 0);
-					const output = t?.predicted_n ?? 0;
-					console.log(
-						`  #${idx + 1}: prompt=${t?.prompt_n ?? 0} cache=${t?.cache_n ?? 0} reading=${read} | output=${output} ${modelLabel}`
-					);
-				}
-			});
-		} else {
-			assistantMessages.forEach((msg, idx) => {
-				const t = msg.timings;
-				const read = (t?.prompt_n ?? 0) + (t?.cache_n ?? 0);
-				const output = t?.predicted_n ?? 0;
-				totalRead += read;
-				totalOutput += output;
-
-				const modelLabel = msg.model ? ` [${msg.model}]` : '';
-				console.log(
-					`  #${idx + 1}: prompt=${t?.prompt_n ?? 0} cache=${t?.cache_n ?? 0} reading=${read} | output=${output} ${modelLabel}`
-				);
-			});
-		}
-
-		// Current request values (what's in KV cache now)
-		const lastMsg = assistantMessages[assistantMessages.length - 1];
-		const currentRead =
-			agenticMessages.length > 0
-				? (agg?.prompt_n ?? 0)
-				: lastMsg?.timings
-					? (lastMsg.timings.prompt_n ?? 0) + (lastMsg.timings.cache_n ?? 0)
-					: 0;
-		const currentOutput =
-			agenticMessages.length > 0 ? (agg?.predicted_n ?? 0) : (lastMsg?.timings?.predicted_n ?? 0);
-
-		console.log('%cCumulative (all responses):', 'font-weight: bold');
-		console.log(`  Reading:   ${formatParameters(totalRead)} tok`);
-		console.log(`  Output:    ${formatParameters(totalOutput)} tok`);
-
-		console.log('%cCurrent (KV cache):', 'font-weight: bold');
-		console.log(`  Reading:   ${formatParameters(currentRead)} tok`);
-		console.log(`  Output:    ${formatParameters(currentOutput)} tok`);
-
-		// Compare with live gauge values (if server state is still available)
-		const serverContext = this.activeProcessingState?.contextUsed ?? 0;
-		console.log('%cLive gauge values:', 'font-weight: bold');
-		console.log(`  contextUsed (server):        ${formatParameters(serverContext)} tok`);
-		console.log(
-			`  contextTotal (model):        ${formatParameters(
-				isRouterMode()
-					? (selectedModelContextSize() ?? 'unknown')
-					: singleModelName()
-						? modelsStore.getModelContextSize(singleModelName()!)
-						: 'unknown'
-			)} tok`
-		);
-
-		if (serverContext > 0 && currentRead > 0) {
-			const diff = serverContext - (currentRead + currentOutput);
-			if (diff !== 0) {
-				console.log(
-					`  Server vs calculated delta: ${formatParameters(diff)} tok`,
-					diff > 0 ? '(system prompt + overhead)' : ''
-				);
-			} else {
-				console.log('  ✓ Server and calculated context match');
-			}
-		}
-
-		console.groupEnd();
+		await debugTokenUsage(this.activeProcessingState, convIdOrUndefined);
 	}
 
 	restoreProcessingStateFromMessages(messages: DatabaseMessage[], conversationId: string): void {
